refactor(authup): give the two listener roles a shared factory of their own
`defineApplicationCommand` builds both `start` and `core`, and its doc
comment describes both, so living in `start.ts` made that file the start
command and the shared factory at once and left `core.ts` importing a
sibling command. It moves to `commands/application.ts`; the two role files
are now three lines each.
