+++ apps/authup/src/commands/index.ts
@@ -5,6 +5,7 @@
  *  view the LICENSE file that was distributed with this source code.
  */
 
+export * from './application.ts';
 export * from './config.ts';
 export * from './console.ts';
 export * from './core.ts';
